YetAnotherDisk/requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.8.0
click==8.1.3
colorama==0.4.5
Django==4.1.1
django-filter==22.1
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.2
